change elementtype

+++ code/test2/2.cpp
@@ -3,10 +3,10 @@
 # include<vector>
 using namespace std;
 long long  ans[100] = {3,6};
-int st[9] ={0,1,1,1,0,1,1,1,0},sub[9];
+long long st[9] ={0,1,1,1,0,1,1,1,0},sub[9];
 
 void sum1(int n){
-    int s = 0;
+    long long s = 0;
     for(int i = 0 ; i < 9 ; i++){
         if((i/3) != (i%3)) s+=st[i];
     }
